% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{as.character.regex}
\alias{as.character.regex}
\title{Convert regex to a character}
\usage{
\method{as.character}{regex}(x, ...)
}
\arguments{
\item{x}{\code{regex} object.}

\item{...}{further args ignored by method.}
}
\description{
Convert regex to a character
}

