% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain_trt}
\alias{brain_trt}
\title{a test-retest demo data}
\format{
A data frame with 62 observations and 364 variables:
}
\source{
\url{https://www.humanconnectome.org/}
}
\usage{
brain_trt
}
\description{
a dataframe containing a test-retest demo data of brain connectivity (seed in PCC area) 
diamonds.
}
\keyword{datasets}
