% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbICC.R
\name{calc_dbICC}
\alias{calc_dbICC}
\title{Distance-based Intraclass Correlation (dbICC)}
\usage{
calc_dbICC(dmax, subID, sortdata = TRUE, return_var = TRUE)
}
\arguments{
\item{dmax}{[n, n]: a distance matrix for n observations}

\item{subID}{[n]: a vector containing the subject IDs}

\item{sortdata}{: a Boolean value - if sort the input data based on subID}

\item{return_var}{: a Boolean value - if return the within and between individual variance (default=TRUE)}
}
\value{
distance-based intraclass correlation (dbICC)
}
\description{
A function for computing the dbICC - multivariate reliability
}
\note{
subID should match with the data (row)
}
\examples{
data <- demo_brain_trt[,5:10]
subID <- as.matrix(demo_brain_trt$subID)
calc_dbICC(dist(data), subID)

}
\references{
Meng Xu et al., 2021. Generalized reliability based on distances. Biometrics.https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7984087
}
\author{
Ting Xu modified from https://github.com/wtagr/dbicc
}
