% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprinting.R
\name{calc_fingerprinting}
\alias{calc_fingerprinting}
\title{calculate the Fingerprinting (Identification Rate)}
\usage{
calc_fingerprinting(dist, sID, method = 1)
}
\arguments{
\item{dist}{[n x n] a distance matrix for n subjects.}

\item{sID}{[n] a vector containing the subject ID for each subject.}

\item{method}{1 (default) or 2, 1: count identification for each subject. 2: count identification for each observation)}
}
\value{
FP [n] a vector containing fringerprinting (identification rate) range 0-1
}
\description{
A function for computing the Identification Rate of a dataset.
}
\author{
Ting Xu
}
