% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2C2.R
\name{calc_i2c2}
\alias{calc_i2c2}
\title{calculate the I2C2}
\usage{
calc_i2c2(data, sID, session)
}
\arguments{
\item{data}{[n x n]: distance matrix.}

\item{sID}{[n]: a subject ID vector}

\item{session}{[n]: a vector indicate session(repetition)}
}
\value{
I2C2
}
\description{
call I2C2 function \code{I2C2::I2C2} from https://rdrr.io/github/neuroconductor/I2C2
}
