% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calc.R
\name{calc_icc_gradient_flow}
\alias{calc_icc_gradient_flow}
\title{calculate the ICC gradient flow (change and normalized change of ICC)}
\usage{
calc_icc_gradient_flow(
  df,
  df.ref,
  name = NULL,
  name.ref = NULL,
  return.input = FALSE
)
}
\arguments{
\item{df}{the target dataframe contains within-individual variation named as sigma2_w and between-individual variation named as sigma2_b}

\item{df.ref}{the reference dataframe}

\item{name}{assign a name of the target icc dataset (default: NULL)}

\item{name.ref}{assign a name of the reference icc dataset (default NULL)}

\item{return.input}{if return the input datasets (target is named as .1 the reference is named as .2)}
}
\value{
df.gradient contains normalized sigma2_w, normalized sigma2_b, delta.icc, normalized theta [0, 2pi), contrast name,
}
\description{
call function \code{gradient_flow} to calculate the ICC gradient flow
}
\references{
A Guide for Quantifying and Optimizing Measurement Reliability for the Study of Individual Differences. doi: https://doi.org/10.1101/2022.01.27.478100
}
\author{
Ting XU
}
