% Generated by roxygen2 (4.0.2): do not edit by hand
\name{capture}
\alias{.}
\alias{..}
\alias{capture}
\alias{capture_group}
\alias{named_capture}
\title{Create a capture group}
\usage{
capture(...)

named_capture(name, ...)

capture_group(name)
}
\arguments{
\item{...}{variables to capture.  Can be \code{\link{shortcuts}}, R
variables, text, or other \pkg{rex} functions.}

\item{name}{of the group.  Unnamed capture groups are numbers starting at 1
in the order they appear in the regular expression.  If two groups have the
same name, the leftmost group is the used in any reference.}
}
\description{
Used to save the matched value within the group for use later in the regular
expression or to extract the values captured.  Types of capture groups.
\describe{
  \item{Unnamed Capture}{- \code{\link{capture}} or
  \code{\link{.}}.  The groups are numbered incrementally starting at one.}
  \item{Named Capture}{- \code{\link{named_capture}} or
  \code{\link{..}}.  The groups are named explicitly by the user.}
}
}
\examples{
# Match paired quotation marks
re <- rex(
  # first quotation mark
  capture(quotes),

  # match all non-matching quotation marks
  zero_or_more(except(capture_group(1))),

  # end quotation mark (matches first)
  capture_group(1)
)
}
\seealso{
\code{\link{group}} for grouping without capturing.  Perl 5 capture
groups \url{http://perldoc.perl.org/perlre.html#Capture-groups}

Other rex: \code{\link{group}}
}

