\name{char}
\alias{char}
\title{Word Character}
\usage{
char(x = NULL)
}
\arguments{
  \item{x}{Either \code{NULL} to indicate any character, or
  a word.}
}
\description{
This function is used to any match any single word
character if called with no arguments, or matches the word
input exactly, in the regular expression.
}
\details{
A word character is any (lowercase or uppercase) letter,
number, or underscore.
}
\seealso{
\code{\link{non_word}}
}

