% Generated by roxygen2 (4.0.2): do not edit by hand
\name{character_class}
\alias{character_class}
\alias{except}
\alias{one_of}
\alias{range}
\title{Create character classes}
\usage{
character_class(x)

one_of(...)

except(...)

range(x, y)
}
\arguments{
\item{...}{\code{\link{shortcuts}}, R variables, text, or other \pkg{rex}
functions.}
}
\description{
There are four different ways you can create a character class.
}
\section{Functions}{
\itemize{
\item \code{}: explicitly define a character class

\item \code{}: specify which characters to include

\item \code{}: specify which characters to exclude

\item \code{}: specify a range of which characters to include
}}
\seealso{
Other rex: \code{\link{.}}, \code{\link{..}},
  \code{\link{capture}}, \code{\link{capture_group}},
  \code{\link{named_capture}}; \code{\link{group}}
}

