% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/escape.R
\name{character_class_escape}
\alias{character_class_escape}
\alias{character_class_escape.character}
\alias{character_class_escape.character_class}
\alias{character_class_escape.default}
\alias{character_class_escape.list}
\alias{character_class_escape.regex}
\title{Character class escapes}
\usage{
character_class_escape(x)

\method{character_class_escape}{regex}(x)

\method{character_class_escape}{character_class}(x)

\method{character_class_escape}{character}(x)

\method{character_class_escape}{list}(x)

\method{character_class_escape}{default}(x)
}
\arguments{
\item{x}{Object to escape.}
}
\description{
Character class escapes
}
\section{Methods (by class)}{
\itemize{
\item \code{regex}: objects are passed through unchanged.

\item \code{character_class}: objects are passed through unchanged.

\item \code{character}: objects properly escaped for character classes.

\item \code{list}: call \code{character_class_escape} on all elements of the list.

\item \code{default}: coerce to \code{character} and \code{character_class_escape}.
}}

