% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{color.brewer.n}
\alias{color.brewer.n}
\title{color.brewer.n
select a subset color from RColorBrewer palettes}
\usage{
color.brewer.n(n.color = NULL, pla.name = "Set2")
}
\arguments{
\item{n.color}{a integer - the number of subset color}

\item{pla.name}{a character - a color palette name (default: 'Set2') from RColorBrewer palettes Check \code{RColorBrewer::display.brewer.all()}}
}
\value{
a Hex color vector
}
\description{
color.brewer.n
select a subset color from RColorBrewer palettes
}
