% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{colormap.gradient.flow}
\alias{colormap.gradient.flow}
\title{colormap of the icc gradient flow
colormap of the icc gradient flow (circular color map)}
\usage{
colormap.gradient.flow()
}
\value{
[360 x 3] a RGB color map
}
\description{
colormap of the icc gradient flow
colormap of the icc gradient flow (circular color map)
}
