% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_brain_trt}
\alias{demo_brain_trt}
\title{a test-retest demo data}
\format{
A data frame with 62 observations and 364 variables:
}
\source{
\url{https://www.humanconnectome.org/}
}
\usage{
demo_brain_trt
}
\description{
a dataframe containing a test-retest demo data of brain connectivity (seed is the posterial cingular cortex)
}
\keyword{datasets}
