% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_delta_var}
\alias{demo_delta_var}
\title{the change of the variation calculate using icc_gradient_flow}
\format{
A data frame with 360 observations and 8 variables:
}
\source{
\url{https://www.humanconnectome.org/}
}
\usage{
demo_delta_var
}
\description{
a dataframe containing the raw and normalized change of the intra- and inter-individual variation (sigma2_w, sigma2_b)
of fMRI-based connectivities with GSR compared to those without GSR 
See data demo_icc_GSR and demo_icc_noGSR
}
\details{
delta.sigma2_b is the change of inter-individual variation
delta.sigma2_w is the change of intra-individual variation
delta.sigma2_b_norm is the relative (normalized) change of inter-individual variation
delta.sigma2_w_norm is the relative (normalized)change of intra-individual variation
}
\keyword{datasets}
