% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_icc_noGSR}
\alias{demo_icc_noGSR}
\title{ICC and variation output demo using lme_ICC_1wayR}
\format{
A data frame with 360 observations and 7 variables:
}
\source{
\url{https://www.humanconnectome.org/}
}
\usage{
demo_icc_noGSR
}
\description{
a dataframe containing ICC, intra- and inter-individual variation of
brain connectivities between the posterial cingular cortex and all 360 brain regions
based on test-retest fMRI data that preprocessed without global signal regression (GSR)
}
\details{
sigma2_b is the inter-individual variation
sigma2_w is the intra-individual variation
}
\keyword{datasets}
