% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminability.R
\name{discr.discr}
\alias{discr.discr}
\title{Discriminability}
\usage{
discr.discr(
  rdf,
  remove_outliers = TRUE,
  thresh = 0,
  output = FALSE,
  all_discr.return = FALSE
)
}
\arguments{
\item{rdf}{[n]: the reliability density function. call \code{discr.rdf}}

\item{remove_outliers}{a boolean indicating whether to ignore subjects with rdf below a certain cutoff (Default: TRUE)}

\item{thresh}{[1]: a numberic value - the threshold, ignore subjects < thred (Default: 0)}

\item{output}{a boolean value indicating whether to ignore output (Default: FALSE)}

\item{all_discr.return}{a boolean value indicating if return the discriminability for each observation (Default: FALSE)}
}
\value{
discr [1]: the discriminability statistic.
}
\description{
A function for computing the discriminability.
}
\author{
Original: Eric Bridgeford and Gregory Kiar; Modified: Ting Xu, return all discriminaiblity for each within-individual repetition
}
