% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calc.R
\name{Discr_and_FP_wraper}
\alias{Discr_and_FP_wraper}
\title{Reliability Explorer wrap-up function - discriminability & fingerprinting}
\usage{
Discr_and_FP_wraper(
  data,
  sID,
  visit,
  method.dist = "euclidean",
  all_discr.return = TRUE,
  method.FP = 1,
  shiny_progress_bar = FALSE
)
}
\arguments{
\item{data}{[n x p] data matrix: n observations x p variables}

\item{sID}{[n] a vector containing subject ID}

\item{visit}{[n] a vector containing the repetition identify (e.g. time1, time2)}

\item{method.dist}{method to calculate the distance matrix (default: "euclidean")}

\item{all_discr.return}{return the discriminability for each observation/repetition}

\item{method.FP}{a numeric value indicating which method used to count the identification rate (default: 1). Check \code{calc_fingerprinting}}

\item{shiny_progress_bar}{a Boolean value - show the Shiny progress bar if this function is called in shiny app (shinybusy library is required)}
}
\value{
a list containing identification rate (fingerprinting) for each variable
}
\description{
A function for computing the fingerprinting for each row variable in data (show progress bar in shiny)
}
\references{
A Guide for Quantifying and Optimizing Measurement Reliability for the Study of Individual Differences. doi: https://doi.org/10.1101/2022.01.27.478100
}
\author{
Ting XU
}
