% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calc.R
\name{discriminability_wraper}
\alias{discriminability_wraper}
\title{Reliability Explorer wrap-up function - discriminability}
\usage{
discriminability_wraper(
  data,
  sID,
  visit,
  method.dist = "euclidean",
  all_discr.return = TRUE,
  shiny_progress_bar = FALSE
)
}
\arguments{
\item{data}{[n x p] data matrix: n observations x p variables}

\item{sID}{[n] a vector containing subject ID}

\item{visit}{[n] a vector containing the repetition identify (e.g. time1, time2)}

\item{method.dist}{[1] a character indicating which method is used to calcualte the distance (Default: euclidean). Check \code{dist}}

\item{all_discr.return}{[1] a boolean value - if return the discriminability for each observation (Default: TRUE)}

\item{shiny_progress_bar}{[1] a boolean value - show the Shiny progress bar if this function is called in shiny app (shinybusy library is required)}
}
\value{
out_list: a list containing discriminability (Discr) and/or discriminability for each repetition/observation (DiscrSub) for each variable
}
\description{
A function for computing the discriminability for each row variable in data (show progress bar in shiny)
}
\references{
A Guide for Quantifying and Optimizing Measurement Reliability for the Study of Individual Differences. doi: https://doi.org/10.1101/2022.01.27.478100
}
\author{
Ting XU
}
