% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_field.R
\name{distance_field_map.df}
\alias{distance_field_map.df}
\title{convert distance matrix to a dataframe}
\usage{
distance_field_map.df(Dmax, subID)
}
\arguments{
\item{Dmax}{[n x n] distance matrix. Check \code{dist}}

\item{subID}{[n] a vector containing subject ID}
}
\value{
df: a dataframe that can be used in \code{plot.distance_field_map}, which contains variables: distance, wD, bD
}
\description{
convert distance matrix to a dataframe for \code{plot.distance_field_map}, which contains variables: distance, wD, bD
}
\references{
A Guide for Quantifying and Optimizing Measurement Reliability for the Study of Individual Differences. doi: https://doi.org/10.1101/2022.01.27.478100
}
\author{
Ting XU
}
