% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape.R
\name{escape}
\alias{escape}
\alias{escape.character}
\alias{escape.character_class}
\alias{escape.default}
\alias{escape.list}
\alias{escape.regex}
\title{Escape characters for a regex}
\usage{
escape(x)

\method{escape}{regex}(x)

\method{escape}{character_class}(x)

\method{escape}{character}(x)

\method{escape}{default}(x)

\method{escape}{list}(x)
}
\arguments{
\item{x}{Object to escape.}
}
\description{
Escape characters for a regex
}
\section{Methods (by class)}{
\itemize{
\item \code{regex}: Objects are simply passed through unchanged.

\item \code{character_class}: Objects are surrounded by braces.

\item \code{character}: Objects are properly escaped for regular expressions.

\item \code{default}: default escape coerces to character and escapes.

\item \code{list}: simply call escape on all elements of the list.
}}

