% Generated by roxygen2 (4.0.2): do not edit by hand
\name{group}
\alias{group}
\title{Create a grouped expression}
\usage{
group(...)
}
\arguments{
\item{...}{variables to capture.  Can be \code{\link{shortcuts}}, R
variables, text, or other \pkg{rex} functions.}
}
\description{
This is similar to \code{\link{capture}} except that it does not store the
value of the group.  This can be used when you want to combine several parts
together but do not need to later reference the grouped value.
}
\seealso{
Other rex: \code{\link{.}}, \code{\link{..}},
  \code{\link{capture}}, \code{\link{capture_group}},
  \code{\link{named_capture}}
}

