% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{hex2rgb}
\alias{hex2rgb}
\title{Hex to RGB conversion
Hex to RGB conversion}
\usage{
hex2rgb(cmap.hex)
}
\arguments{
\item{cmap.hex}{a Hex color vector}
}
\value{
[n x 3] a RGB color matrix
}
\description{
Hex to RGB conversion
Hex to RGB conversion
}
