% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calc.R
\name{icc_gradient_flow}
\alias{icc_gradient_flow}
\title{gradient flow function - calculate the normalized change of ICC (sigma2 and theta)}
\usage{
icc_gradient_flow(x1, y1, x0, y0)
}
\arguments{
\item{x1}{within-individual variation of the target (end point)}

\item{y1}{between-individual variation of the target (end point)}

\item{x0}{within-individual variation of the reference (start point)}

\item{y0}{between-individual variation of the reference (start point)}
}
\value{
df: a dataframe contains normalized sigma2_w, normalized sigma2_b, normalized theta in  [0, 2pi)
}
\description{
calculate the normalied sigma2_w, normalized sigma2_b and normalized theta in [0, 2pi)
}
\references{
A Guide for Quantifying and Optimizing Measurement Reliability for the Study of Individual Differences. doi: https://doi.org/10.1101/2022.01.27.478100
}
\author{
Ting XU
}
