\name{if_next_is}
\alias{if_next_is}
\title{Lookarounds}
\usage{
if_next_is(x)
}
\arguments{
  \item{x}{A regex pattern.}

  \item{y}{A regex pattern.}
}
\description{
These functions provide an interface to perl lookarounds.
}
\details{
Special binary functions are used to infer an ordering,
since often you might wish to match a word / set of
characters conditional on the start and end of that word.
}

