% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_ICC.R
\name{lme_ICC_2wayR}
\alias{lme_ICC_2wayR}
\title{ICC (ICC2) calculation using 2-way Random model}
\usage{
lme_ICC_2wayR(data, subID, session, cov = NULL, shiny_progress_bar = FALSE)
}
\arguments{
\item{data}{[n, p]: a data matrix for n observations and p variables.}

\item{subID}{[n]: a vector containing the subject IDs for each subject.}

\item{session}{[n]: a vector containing the session (run, subsets, site, etc.)}

\item{cov}{[n, m]: a covariance matrix for n observations and m variables}

\item{shiny_progress_bar}{[1] a Boolean value - show the Shiny progress bar if this function is called in shiny app (library(shinybusy) is required)}
}
\value{
ICC: output vector contains the following elements:

ICC_a: 2-way random model, Agreement, single raters, defined ad ICC(2,1)

ICC_c: 2-way random model, Consistency, single raters

ICCk_a: 2-way random model, Agreement, multiple raters, defined ad ICC(2,k)

ICCk_c:  2-way random model, Consistency, multiple raters

sigma2_b: sigma^2(between-sub), estimated between-subject variation

sigma2_s: sigma^2(session),  estimated between-session variation

sigma2_r: sigma^2(residual),  estimated within-subject variation

var(data) - variation of the dependent variable(s)
}
\description{
Calculate Intra-class correlation (ICC) and variation using Linear Mixed Model (LMM) 
with the Restricted Maximum Likelihood (ReML) estimation.
}
\details{
Model: 2-way random model using lme4 package. Base model: y ~ 1 + (1|session) + (1|subID)

Output: ICC agreement or consistency, single and multiple raters.
}
\references{
A Guide for Quantifying and Optimizing Measurement Reliability for the Study of Individual Differences. doi: https://doi.org/10.1101/2022.01.27.478100
}
\author{
Ting Xu
}
