% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lookarounds}
\alias{lookarounds}
\title{Lookarounds}
\arguments{
\item{x}{A regex pattern.}

\item{y}{A regex pattern.}
}
\description{
Lookarounds
}
\details{
These functions provide an interface to perl lookarounds.

Special binary functions are used to infer an ordering, since often you
might wish to match a word / set of characters conditional on the start
and end of that word.
}

