% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{not}
\alias{not}
\title{Do not match}
\usage{
not(..., type = c("greedy", "lazy", "possessive"))
}
\arguments{
\item{...}{\code{\link{shortcuts}}, R variables, text, or other \pkg{rex}
functions.}
}
\description{
Do not match
}
\seealso{
Other rex: \code{\link{.}}, \code{\link{capture}},
  \code{\link{capture_group}};
  \code{\link{\%if_next_is\%}},
  \code{\link{\%if_next_isnt\%}},
  \code{\link{\%if_prev_is\%}},
  \code{\link{\%if_prev_isnt\%}},
  \code{\link{lookarounds}}; \code{\link{at_least}},
  \code{\link{at_most}}, \code{\link{between}},
  \code{\link{counts}}, \code{\link{n}},
  \code{\link{n_times}}; \code{\link{character_class}},
  \code{\link{except}}, \code{\link{exclude_range}},
  \code{\link{none_of}}, \code{\link{one_of}},
  \code{\link{range}}; \code{\link{group}};
  \code{\link{rex}}, \code{\link{rex_}}
}

