% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or.R
\name{\%or\%}
\alias{\%or\%}
\alias{or}
\title{Or}
\usage{
x \%or\% y

or(...)
}
\arguments{
\item{x}{A string.}

\item{y}{A string.}

\item{...}{\code{\link{shortcuts}}, R variables, text, or other \pkg{rex}
functions.}
}
\description{
The special binary function \code{\%or\%} can be used to specify a set
of optional matches.
}
\seealso{
Other rex: 
\code{\link{capture}()},
\code{\link{character_class}()},
\code{\link{counts}},
\code{\link{group}()},
\code{\link{lookarounds}},
\code{\link{not}()},
\code{\link{rex}()},
\code{\link{shortcuts}},
\code{\link{wildcards}}
}
\concept{rex}
