% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{print.regex}
\alias{print.regex}
\alias{regex}
\title{Regular Expression}
\usage{
\method{print}{regex}(x, ...)

regex(x)
}
\arguments{
\item{x}{\code{regex} object to be printed}

\item{...}{further args ignored by method}
}
\description{
Specify an explicit regular expression.  This expression must already be
escaped.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Object printing
}}

