% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{register_shortcuts}
\alias{register_shortcuts}
\title{Register the Rex shortcuts}
\usage{
register_shortcuts(pkg_name)
}
\arguments{
\item{pkg_name}{the package to register the shortcuts in}
}
\description{
If you are using rex in another package you need to call this function to
register all of the rex shortcuts so that spurious NOTEs about global
variables being generated during R CMD check.
}

