% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/aaa.R
\name{reregister_functions}
\alias{reregister_functions}
\title{Re-register the rex functions}
\usage{
reregister_functions(env)
}
\arguments{
\item{env}{the environment to take the new definitions from.}
}
\description{
Use if you need to re-define the functions after package load.
}

