% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{rex_mode}
\alias{rex_mode}
\title{Toggles \pkg{rex} mode.}
\usage{
rex_mode()
}
\description{
While within rex mode, functions used within the \code{\link{rex}} function
are attached, so one can get e.g. auto-completion within editors.
}

