% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_field.R
\name{rex_plot.distance}
\alias{rex_plot.distance}
\title{plot distance matrix}
\usage{
rex_plot.distance(
  Dmax,
  gtitle = "",
  color = c("white", "red"),
  show.diag = TRUE
)
}
\arguments{
\item{Dmax}{Distance matrix}

\item{gtitle}{plot title (default: NULL)}

\item{color}{palette name from ColorBrewer, or a color array, e.g. c("blue", "red"). By default, color=c("white", "red)
Check out the what color palette available
\code{library(ColorBrewer); display.brewer.all()}}

\item{show.diag}{If true, assign 0 to diagonal matrix}
}
\value{
ggplot object of the distance matrix
}
\description{
plot distance matrix - ggplot based
}
\references{
A Guide for Quantifying and Optimizing Measurement Reliability for the Study of Individual Differences. doi: https://doi.org/10.1101/2022.01.27.478100
}
\author{
Ting XU
}
