% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variation_field.R
\name{rex_plot.icc.gradient.hist}
\alias{rex_plot.icc.gradient.hist}
\title{plot function - plot the histogram of ICC gradient flow, the normalized change of the variation}
\usage{
rex_plot.icc.gradient.hist(
  df.plot,
  plot.title = NULL,
  nbins = 40,
  show.ytick = FALSE,
  color.histline = NULL,
  show.ref.line = TRUE,
  show.ref.legend = TRUE
)
}
\arguments{
\item{df.plot}{dataframe for plot containing delta.sigma2_w, delta.sigma2_b}

\item{plot.title}{title of the plot}

\item{nbins}{a integer value - the number of the histogram}

\item{show.ytick}{a Boolean value - if showing the ytick. Default: FALSE}

\item{color.histline}{a color name or Hex code - the color of the histgram line. Default: NULL}

\item{show.ref.line}{a Boolean value - if show the reference lines. Default: TRUE}

\item{show.ref.legend}{a Boolean value - if show the reference legend Default: TRUE}
}
\value{
ggplot plot object
}
\description{
histgram of the ICC gradient (x-axis: relative change of the within-individual variation, y-axis: relative change of the between-individual variation).
}
\note{
If plot.point=TRUE, the bin plot is disabled.
}
\references{
A Guide for Quantifying and Optimizing Measurement Reliability for the Study of Individual Differences. doi: https://doi.org/10.1101/2022.01.27.478100
}
\author{
Ting Xu
}
