% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variation_field.R
\name{rex_plot.icc.gradient.norm}
\alias{rex_plot.icc.gradient.norm}
\title{plot function - plot the ICC gradient flow, the normalized change of the variation}
\usage{
rex_plot.icc.gradient.norm(
  df.plot,
  plot.title = NULL,
  axis.min = NULL,
  axis.max = NULL,
  plot.percent = c(0, 1),
  plot.point = TRUE,
  size.point = 2,
  color.point.gradient = TRUE,
  color.point.fill = "red",
  color.point.border = NULL,
  plot.bin = FALSE,
  nbins = 30,
  color.bin = "red",
  ColorBrewer.direction = 1,
  show.contour = TRUE,
  nbins.contour = 6,
  color.contour = "grey30",
  show.ref.line = TRUE,
  show.ref.legend = TRUE
)
}
\arguments{
\item{df.plot}{dataframe for plot containing delta.sigma2_w, delta.sigma2_b}

\item{plot.title}{title of the plot}

\item{axis.min}{the minimum value of x- and y-axis}

\item{axis.max}{the maximum value of x- and y-axis}

\item{plot.percent}{a vector - the percentile cut of the data that are shown in the plot. Default: c(0,1)}

\item{plot.point}{a Boolean value - if plot each of the input as a dot. Default: TRUE}

\item{size.point}{a integer value - the size of the dots if plot.point = TRUE}

\item{color.point.gradient}{a Boolean value - if use the standard gradient color}

\item{color.point.fill}{a color name or Hex code - the color to fill the dots}

\item{color.point.border}{a color name or Hex code - the color of the dots bolder}

\item{plot.bin}{a Boolean value - heatmap of 2d bin (call ggplot stat_bin_2d ). If plot.density=TRUE, plot.bin is turned off}

\item{nbins}{a integer value - the number of the 2d bins of the heatmap (call ggplot stat_bin_2d)}

\item{color.bin}{- a color/palette name or vector, a color array, in color name or Hex Code, or palette name for the density plot. Default: "red".
a color name or a Hex code "##ff0000" or "red". It will be filled by c("white", "red"). 
a color array in color name or Hex code, e.g. c("white", "blue", "red").
palette name from ColorBrewer. check what is available RColorBrewer::display.brewer.all()
Note, if only one color provided, e.g. "red", then it is filled by c("white", "red").}

\item{ColorBrewer.direction}{1 or -1 - if a ColorBrewer color palette is assigned to color, this parameter controls the direction}

\item{show.contour}{a Boolean value - if show the contour line of the density plot}

\item{nbins.contour}{a integer value - the number of the contour lines for the density plot. Default: 6}

\item{color.contour}{a color name or Hex code - the color of the contour line. Default: "grey30"}

\item{show.ref.line}{a Boolean value - if show the reference lines. Default: TRUE}

\item{show.ref.legend}{a Boolean value - if show the reference legend}
}
\value{
ggplot plot object
}
\description{
variation field (x-axis: relative change of the within-individual variation, y-axis: relative change of the between-individual variation).
}
\note{
If plot.point=TRUE, the bin plot is disabled.
}
\references{
A Guide for Quantifying and Optimizing Measurement Reliability for the Study of Individual Differences. doi: https://doi.org/10.1101/2022.01.27.478100
}
\author{
Ting Xu
}
