% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_field.R
\name{rex_plot.sub_distance}
\alias{rex_plot.sub_distance}
\title{plot distance matrix, ordered by subID and visit}
\usage{
rex_plot.sub_distance(Dmax, subID, visit, gtitle = "")
}
\arguments{
\item{Dmax}{[n x n] a matrix, Distance matrix}

\item{subID}{[n] a vector, subject ID}

\item{visit}{[n] a vector, visit (e.g. time1, time2)}

\item{gtitle}{plot title (default: NULL)}
}
\value{
pdist: ggplot object of the distance matrix
}
\description{
plot distance matrix, ordered by subID and visit
}
\note{
subID and visit should match with the row/column of the Dmax. The color of the plot is white to red
}
\references{
A Guide for Quantifying and Optimizing Measurement Reliability for the Study of Individual Differences. doi: https://doi.org/10.1101/2022.01.27.478100
}
\author{
Ting XU
}
