% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variation_field.R
\name{rex_plot.var.field}
\alias{rex_plot.var.field}
\title{plot function - plot the variation field map}
\usage{
rex_plot.var.field(
  df.plot,
  plot.title = NULL,
  axis.min = NULL,
  axis.max = NULL,
  plot.percent = c(0, 1),
  plot.point = TRUE,
  size.point = 2,
  color.point.fill = "red",
  color.point.border = "grey80",
  plot.bin = FALSE,
  nbins = 30,
  plot.density = TRUE,
  nbins.contour = 6,
  alpha.density = 0.5,
  color = "red",
  ColorBrewer.direction = 1,
  show.contour = FALSE,
  color.contour = "grey30",
  show.icc_slope = TRUE,
  color.icc_slope = "grey50"
)
}
\arguments{
\item{df.plot}{dataframe for plot containing sigma2_w, sigma2_b}

\item{plot.title}{title of the plot}

\item{axis.min}{the minimum value of x- and y-axis}

\item{axis.max}{the maximum value of x- and y-axis}

\item{plot.percent}{a vector - the percentile cut of the data that are shown in the plot. Default: c(0,1)}

\item{plot.point}{a Boolean value - if plot each of the input as a dot. Default: TRUE}

\item{size.point}{a integer value - the size of the dots if plot.point = TRUE}

\item{color.point.fill}{a Hex code or Color name that can be read in ggplot - the color of the dots}

\item{color.point.border}{a Hex code or Color name that can be read in ggplot - the color of the dots' border}

\item{plot.bin}{a Boolean value - heatmap of 2d bin (call ggplot stat_bin_2d ). If plot.density=TRUE, plot.bin is turned off}

\item{nbins}{a integer value - the number of the 2d bins of the heatmap (call ggplot stat_bin_2d)}

\item{plot.density}{a Boolean value - if plot the 2d density plot (call ggplot stat_density_2d). Default:TRUE}

\item{nbins.contour}{a integer value - the number of the contour lines for the density plot. Default: 6}

\item{alpha.density}{a numeric value - the alpha value if make the density plot transparent, range 0-1. Default: 0.5}

\item{color}{- a color/palette name or vector, a color array, in color name or Hex Code, or palette name for the density plot. Default: "red".
a color name or a Hex code "##ff0000" or "red". It will be filled by c("white", "red"). 
a color array in color name or Hex code, e.g. c("white", "blue", "red").
palette name from ColorBrewer. check what is available RColorBrewer::display.brewer.all()
Note, if only one color provided, e.g. "red", then it is filled by c("white", "red").}

\item{ColorBrewer.direction}{1 or -1 - if a ColorBrewer color palette is assigned to color, this parameter controls the direction}

\item{show.contour}{a Boolean value - if show the contour line of the density plot}

\item{color.contour}{a color name or Hex code - the color of the contour line. Default: "grey30"}

\item{show.icc_slope}{a Boolean value - if show the icc reference slope lines. Default: TRUE}

\item{color.icc_slope}{a color value - the color of the icc reference slope lines. Default: "grey50"}
}
\value{
ggplot plot object
}
\description{
variation field (x-axis: within-individual variation, y-axis: between-individual variation).
}
\references{
A Guide for Quantifying and Optimizing Measurement Reliability for the Study of Individual Differences. doi: https://doi.org/10.1101/2022.01.27.478100
}
\author{
Ting Xu
}
