% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{rgb2hex}
\alias{rgb2hex}
\title{RGB to Hex conversion
RGB to Hex conversion}
\usage{
rgb2hex(cmap.rgb)
}
\arguments{
\item{cmap.rgb}{[n x 3] a RGB color matrix}
}
\value{
a Hex color vector
}
\description{
RGB to Hex conversion
RGB to Hex conversion
}
