% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\docType{data}
\name{shortcuts}
\alias{shortcuts}
\title{Shortcuts}
\format{\preformatted{dot - \\.
any - .
something - .+
anything - .*
start - ^
end - $
boundary - \\b
non_boundary - \\B
alnum - [:alnum:]
alpha - [:alpha:]
letter - [:alpha:]
blank - [:blank:]
cntrl - [:cntrl:]
digit - [:digit:]
number - [:digit:]
graph - [:graph:]
lower - [:lower:]
print - [:print:]
punct - [:punct:]
space - [:space:]
upper - [:upper:]
xdigit - [:xdigit:]
newline - \\R
single_quote - '
double_quote - "
quote - '"
alnums - [[:alnum:]]+
alphas - [[:alpha:]]+
letters - [[:alpha:]]+
blanks - [[:blank:]]+
cntrls - [[:cntrl:]]+
digits - [[:digit:]]+
numbers - [[:digit:]]+
graphs - [[:graph:]]+
lowers - [[:lower:]]+
prints - [[:print:]]+
puncts - [[:punct:]]+
spaces - [[:space:]]+
uppers - [[:upper:]]+
xdigits - [[:xdigit:]]+
newlines - \\R+
single_quotes - [']+
double_quotes - ["]+
quotes - ['"]+
any_alnums - [[:alnum:]]*
any_alphas - [[:alpha:]]*
any_letters - [[:alpha:]]*
any_blanks - [[:blank:]]*
any_cntrls - [[:cntrl:]]*
any_digits - [[:digit:]]*
any_numbers - [[:digit:]]*
any_graphs - [[:graph:]]*
any_lowers - [[:lower:]]*
any_prints - [[:print:]]*
any_puncts - [[:punct:]]*
any_spaces - [[:space:]]*
any_uppers - [[:upper:]]*
any_xdigits - [[:xdigit:]]*
any_newlines - \\R*
any_single_quotes - [']*
any_double_quotes - ["]*
any_quotes - ['"]*
non_alnum - ^[:alnum:]
non_alpha - ^[:alpha:]
non_letter - ^[:alpha:]
non_blank - ^[:blank:]
non_cntrl - ^[:cntrl:]
non_digit - ^[:digit:]
non_number - ^[:digit:]
non_graph - ^[:graph:]
non_lower - ^[:lower:]
non_print - ^[:print:]
non_punct - ^[:punct:]
non_space - ^[:space:]
non_upper - ^[:upper:]
non_xdigit - ^[:xdigit:]
non_newline - ^\\R
non_single_quote - ^'
non_double_quote - ^"
non_quote - ^'"
non_alnums - [^[:alnum:]]+
non_alphas - [^[:alpha:]]+
non_letters - [^[:alpha:]]+
non_blanks - [^[:blank:]]+
non_cntrls - [^[:cntrl:]]+
non_digits - [^[:digit:]]+
non_numbers - [^[:digit:]]+
non_graphs - [^[:graph:]]+
non_lowers - [^[:lower:]]+
non_prints - [^[:print:]]+
non_puncts - [^[:punct:]]+
non_spaces - [^[:space:]]+
non_uppers - [^[:upper:]]+
non_xdigits - [^[:xdigit:]]+
non_newlines - ^\\R+
non_single_quotes - [^']+
non_double_quotes - [^"]+
non_quotes - [^'"]+
any_non_alnums - [^[:alnum:]]*
any_non_alphas - [^[:alpha:]]*
any_non_letters - [^[:alpha:]]*
any_non_blanks - [^[:blank:]]*
any_non_cntrls - [^[:cntrl:]]*
any_non_digits - [^[:digit:]]*
any_non_numbers - [^[:digit:]]*
any_non_graphs - [^[:graph:]]*
any_non_lowers - [^[:lower:]]*
any_non_prints - [^[:print:]]*
any_non_puncts - [^[:punct:]]*
any_non_spaces - [^[:space:]]*
any_non_uppers - [^[:upper:]]*
any_non_xdigits - [^[:xdigit:]]*
any_non_newlines - ^\\R*
any_non_single_quotes - [^']*
any_non_double_quotes - [^"]*
any_non_quotes - [^'"]*}}
\usage{
shortcuts
}
\description{
Commonly used character classes and regular expressions.  These shortcuts
are substituted inside \code{rex} calls.
}
\details{
\code{names(shortcuts)} will give you the full list of available shortcuts.
}
\seealso{
Other rex: \code{\link{\%or\%}}, \code{\link{capture}},
  \code{\link{character_class}}, \code{\link{counts}},
  \code{\link{group}}, \code{\link{lookarounds}},
  \code{\link{not}}, \code{\link{rex}},
  \code{\link{wildcards}}
}
\keyword{datasets}

