% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\docType{data}
\name{single_shortcuts}
\alias{single_shortcuts}
\title{Single shortcuts}
\format{An object of class \code{list} of length 18.}
\usage{
single_shortcuts
}
\description{
Each of these shortcuts has both a plural (-s) and inverse (non_) form.
}
\keyword{datasets}

