% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cmdline.R
\name{unpack_var}
\alias{unpack_var}
\title{unpack the python-style arguments}
\usage{
unpack_var(var_input)
}
\arguments{
\item{var_input}{- a character input specify the variable names separate by ',' or index(indices). For example
5:10              will take the 5-th to the 10-th columns 
3,4,5:10          will take the 3rd, 4-th and 5-th to the 10-th columns 
var1,var2,var3    will take 3 variables named var1, var2, and var3}
}
\value{
numeric vector or character list
}
\description{
convert the python-style variable name(s) or indices read in from the command line using \code{optparse} package into R readable arguments
}
