% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex.R
\name{as.regex}
\alias{as.regex}
\alias{as.regex.default}
\title{Coerce objects to a \code{\link{regex}}.}
\usage{
as.regex(x, ...)

\method{as.regex}{default}(x, ...)
}
\arguments{
\item{x}{Object to coerce to \code{\link{regex}}.}

\item{...}{further arguments passed to methods.}
}
\description{
Coerce objects to a \code{\link{regex}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Simply escape the Object.
}}

