% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{character_class}
\alias{character_class}
\alias{except}
\alias{exclude_range}
\alias{none_of}
\alias{one_of}
\alias{range}
\title{Create character classes}
\usage{
character_class(x)

one_of(...)

none_of(...)

range(start, end)

exclude_range(start, end)
}
\arguments{
\item{x}{text to include in the character class (must be escaped manually)}

\item{...}{\code{\link{shortcuts}}, R variables, text, or other \pkg{rex}
functions.}

\item{start}{beginning of character class}

\item{end}{end of character class}
}
\description{
There are multiple ways you can create a character class.
}
\section{Functions}{
\itemize{
\item \code{character_class}: explicitly define a character class

\item \code{one_of}: specify which characters to include

\item \code{none_of}: specify which characters to exclude

\item \code{range}: specify a range of which characters to include

\item \code{exclude_range}: specify a range of which characters to exclude
}}
\examples{
# grey = gray
re <- rex("gr", one_of("a", "e"), "y")
grepl(re, c("grey", "gray")) # TRUE TRUE

# Match non-vowels
re <- rex(none_of("a", "e", "i", "o", "u"))
# They can also be in the same string
re <- rex(none_of("aeiou"))
grepl(re, c("k", "l", "e")) # TRUE TRUE FALSE

# Match range
re <- rex(range("a", "e"))
grepl(re, c("b", "d", "f")) # TRUE TRUE FALSE

# Explicit creation
re <- rex(character_class("abcd\\\\["))
grepl(re, c("a", "d", "[", "]")) # TRUE TRUE TRUE FALSE
}
\seealso{
Other rex: \code{\link{.}}, \code{\link{..}},
  \code{\link{capture}}, \code{\link{capture_group}},
  \code{\link{named_capture}};
  \code{\link{\%if_next_is\%}},
  \code{\link{\%if_next_isnt\%}},
  \code{\link{\%if_prev_is\%}},
  \code{\link{\%if_prev_isnt\%}},
  \code{\link{lookarounds}}; \code{\link{at_least}},
  \code{\link{at_most}}, \code{\link{between}},
  \code{\link{counts}}, \code{\link{n}},
  \code{\link{n_times}}; \code{\link{group}}
}

