% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{character_class}
\alias{character_class}
\alias{except}
\alias{exclude_range}
\alias{none_of}
\alias{one_of}
\alias{range}
\title{Create character classes}
\usage{
character_class(x)

one_of(...)

none_of(...)

range(x, y)

exclude_range(x, y)
}
\arguments{
\item{...}{\code{\link{shortcuts}}, R variables, text, or other \pkg{rex}
functions.}
}
\description{
There are four different ways you can create a character class.
}
\section{Functions}{
\itemize{
\item \code{character_class}: explicitly define a character class

\item \code{one_of}: specify which characters to include

\item \code{none_of}: specify which characters to exclude

\item \code{range}: specify a range of which characters to include

\item \code{exclude_range}: specify a range of which characters to exclude
}}
\seealso{
Other rex: \code{\link{.}}, \code{\link{..}},
  \code{\link{capture}}, \code{\link{capture_group}},
  \code{\link{named_capture}}; \code{\link{group}}
}

