% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/or.R
\name{\%or\%}
\alias{\%or\%}
\alias{or}
\title{Or}
\usage{
x \%or\% y

or(...)
}
\arguments{
\item{x}{A string.}

\item{y}{A string.}

\item{...}{\code{\link{shortcuts}}, R variables, text, or other \pkg{rex}
functions.}
}
\description{
The special binary function \code{\%or\%} can be used to specify a set
of optional matches.

describeIn or regular function can also be used, useful for more than 2 arguments.
}
\seealso{
Other rex: \code{\link{.}}, \code{\link{capture}},
  \code{\link{capture_group}}; \code{\link{:}},
  \code{\link{any_of}}, \code{\link{character_class}},
  \code{\link{except}}, \code{\link{except_any_of}},
  \code{\link{except_some_of}},
  \code{\link{exclude_range}}, \code{\link{none_of}},
  \code{\link{one_of}}, \code{\link{range}},
  \code{\link{some_of}}; \code{\link{\%if_next_is\%}},
  \code{\link{\%if_next_isnt\%}},
  \code{\link{\%if_prev_is\%}},
  \code{\link{\%if_prev_isnt\%}},
  \code{\link{lookarounds}}; \code{\link{at_least}},
  \code{\link{at_most}}, \code{\link{between}},
  \code{\link{counts}}, \code{\link{n}},
  \code{\link{n_times}}; \code{\link{group}};
  \code{\link{maybe}}, \code{\link{one_or_more}},
  \code{\link{wildcards}}, \code{\link{zero_or_more}},
  \code{\link{zero_or_one}}; \code{\link{not}};
  \code{\link{rex}}, \code{\link{rex_}};
  \code{\link{shortcuts}}
}

