% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{\%or\%}
\alias{\%or\%}
\alias{or}
\title{Or}
\usage{
x \%or\% y

or(...)
}
\arguments{
\item{x}{A string.}

\item{y}{A string.}

\item{...}{\code{\link{shortcuts}}, R variables, text, or other \pkg{rex}
functions.}
}
\description{
The special binary function \code{\%or\%} can be used to specify a set
of optional matches.

describeIn or regular function can also be used, useful for more than 2 arguments.
}

