% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{re_matches}
\alias{re_matches}
\title{Match function}
\usage{
re_matches(data, pattern)
}
\arguments{
\item{data}{character vector to match against}

\item{pattern}{regular expression to use for matching}
}
\value{
if no captures, returns a logical vector the same length as the
input character vector specifying if the relevant value matched or not.  If
there are captures in the regular expression, returns a data.frame with a
column for each capture group.
}
\description{
Match function
}
\examples{
string = c("this is a", "test string")
re_matches(string, rex("test")) # FALSE FALSE

# named capture
re_matches(string, rex(capture(alphas, name = "first_word"), space,
  capture(alphas, name = "second_word")))
#   first_word second_word
# 1       this          is
# 2       test      string

# capture returns NA when it fails to match
re_matches(string, rex(capture("test")))
#      1
# 1 test
# 2 <NA>
}

