% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/rex.R
\name{as.character.regex}
\alias{as.character.regex}
\alias{print.regex}
\alias{regex}
\title{Regular Expression}
\usage{
\method{as.character}{regex}(x, ...)

\method{print}{regex}(x, ...)

regex(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{further arguments}
}
\description{
Specify an explicit regular expression.  This expression must already be
escaped.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.character}: coerce regex object to a character

\item \code{print}: Print regex object
}}
\seealso{
\code{\link{as.regex}} to coerce to a regex object.
}

