% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex.R
\name{rex}
\alias{rex}
\alias{rex_}
\title{Generate a regular expression.}
\usage{
rex(..., env = parent.frame())
}
\arguments{
\item{...}{\code{\link{shortcuts}}, R variables, text, or other \pkg{rex}
functions.}

\item{env}{environment to evaluate the rex expression in.}
}
\description{
Generate a regular expression.
}
\seealso{
Other rex: \code{\link{\%or\%}}, \code{\link{capture}},
  \code{\link{character_class}}, \code{\link{counts}},
  \code{\link{group}}, \code{\link{lookarounds}},
  \code{\link{not}}, \code{\link{shortcuts}},
  \code{\link{wildcards}}
}

