% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{rex_mode}
\alias{rex_mode}
\title{Rex Mode}
\usage{
rex_mode()
}
\description{
\code{rex_mode()} switch \pkg{rex} mode on and off. While within
rex mode, functions used within the \code{\link{rex}} function are
attached, and so one can get e.g. auto-completion within editors.
}

