% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\docType{data}
\name{shortcuts}
\alias{shortcuts}
\title{Shortcuts}
\format{An object of class \code{list} of length 116.}
\usage{
shortcuts
}
\description{
Commonly used character classes and regular expressions.  These shortcuts
are substituted inside \code{rex} calls.
}
\details{
\code{names(shortcuts)} will give you the full list of available shortcuts.
}
\seealso{
Other rex: \code{\link{\%or\%}}, \code{\link{capture}},
  \code{\link{character_class}}, \code{\link{counts}},
  \code{\link{group}}, \code{\link{lookarounds}},
  \code{\link{not}}, \code{\link{rex}},
  \code{\link{wildcards}}
}
\keyword{datasets}

