% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{data}
\name{single_shortcuts}
\alias{single_shortcuts}
\title{Single shortcuts}
\format{\preformatted{List of 17
 $ alnum       :Classes 'character_class', 'regex'  chr "[:alnum:]"
 $ alpha       :Classes 'character_class', 'regex'  chr "[:alpha:]"
 $ letter      :Classes 'character_class', 'regex'  chr "[:alpha:]"
 $ blank       :Classes 'character_class', 'regex'  chr "[:blank:]"
 $ cntrl       :Classes 'character_class', 'regex'  chr "[:cntrl:]"
 $ digit       :Classes 'character_class', 'regex'  chr "[:digit:]"
 $ number      :Classes 'character_class', 'regex'  chr "[:digit:]"
 $ graph       :Classes 'character_class', 'regex'  chr "[:graph:]"
 $ lower       :Classes 'character_class', 'regex'  chr "[:lower:]"
 $ print       :Classes 'character_class', 'regex'  chr "[:print:]"
 $ punct       :Classes 'character_class', 'regex'  chr "[:punct:]"
 $ space       :Classes 'character_class', 'regex'  chr "[:space:]"
 $ upper       :Classes 'character_class', 'regex'  chr "[:upper:]"
 $ xdigit      :Classes 'character_class', 'regex'  chr "[:xdigit:]"
 $ single_quote:Classes 'character_class', 'regex'  chr "'"
 $ double_quote:Classes 'character_class', 'regex'  chr "\\""
 $ quote       :Classes 'character_class', 'regex'  chr "'\\""
}}
\usage{
single_shortcuts
}
\description{
Each of these shortcuts has both a plural (-s) and inverse (non_) form.
}
\keyword{datasets}

