% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\docType{data}
\name{single_shortcuts}
\alias{single_shortcuts}
\title{Single shortcuts}
\format{\preformatted{alnum - [:alnum:]
alpha - [:alpha:]
letter - [:alpha:]
blank - [:blank:]
cntrl - [:cntrl:]
digit - [:digit:]
number - [:digit:]
graph - [:graph:]
lower - [:lower:]
print - [:print:]
punct - [:punct:]
space - [:space:]
upper - [:upper:]
xdigit - [:xdigit:]
newline - \\R
single_quote - '
double_quote - "
quote - '"}}
\usage{
single_shortcuts
}
\description{
Each of these shortcuts has both a plural (-s) and inverse (non_) form.
}
\keyword{datasets}

